+++ NEWS.md
@@ -1,5 +1,12 @@
 # teal.code 0.4.1.9000
 
+# New features
+
+* The `@code` field in the `qenv` class now holds `character` rather than `expression`.
+* `get_code()` is extended by `names` parameter that allows to extract the code just for a
+specific object.
+* Introduced `# @effect` comment-tag in `code` argument of `eval_code` and `new_qenv` functions to support code relationship detection. See more in functions documentation.
+
 # teal.code 0.4.1
 
 ### Miscellaneous